20260114b Creating new button for watering activity only. New structure created for that. Minor cosmetic changes.
