+++ components/configuration-tabs/address.html
@@ -28,7 +28,7 @@
                 <button class="btn btn-small btn-edit" id="habitat-edit-btn">
                     EDIT
                 </button>
-                <button class="btn btn-small btn-delete" id="habitat-delete-btn">
+                <button class="btn-small btn-delete" id="habitat-delete-btn">
                     DELETE
                 </button>
             </div>
