+++ components/configuration-tabs/habitat.html
@@ -25,10 +25,10 @@
         <div class="card-title">
             <h1>Habitat Details</h1>
             <div class="card-title-btn">
-                <button class="btn btn-small btn-edit" id="habitat-edit-btn">
+                <button class="btn-small btn-edit" id="habitat-edit-btn">
                     EDIT
                 </button>
-                <button class="btn btn-small btn-delete" id="habitat-delete-btn">
+                <button class="btn-small btn-delete" id="habitat-delete-btn">
                     DELETE
                 </button>
             </div>
