+++ components/modals/activity/watering.html
@@ -1,11 +1,9 @@
 <!-- New Activity Form -->
 <form id="new-activity-form">
     <div class="form-group">
-        <label for="activity-type">Activity Type *</label>
-        <select id="activity-type" name="activity_type" required>
-            <option value="">Select activity type...</option>
-            <!-- Options will be loaded dynamically -->
-        </select>
+        <label for="activity-type">Activity Type</label>
+        <div class="static-field">Watering</div>
+        <input type="hidden" name="activity_type" value="watering">
     </div>
     
     <div class="form-group">
@@ -27,21 +25,20 @@
 
         <div class="form-grid">
             <div class="form-group">
-                <label for="activity-quantity">Quantity</label>
-                <input type="number" id="activity-quantity" name="quantifier" step="0.01" placeholder="e.g., 250">
+                <label for="activity-quantity">Level *</label>
+                <select id="activity-quantity" name="quantifier" required>
+                    <option value="">Select watering level...</option>
+                    <option value="25">MINIMAL (25)</option>
+                    <option value="50">HALF (50)</option>
+                    <option value="75">ALMOST FULL (75)</option>
+                    <option value="100">FULL (100)</option>
+                </select>
             </div>
-            
+
             <div class="form-group">
                 <label for="activity-unit">Unit</label>
-                <select id="activity-unit" name="unit">
-                    <option value="">Select unit...</option>
-                    <option value="%">Percentage (%)</option>
-                    <option value="ml">Milliliters (ml)</option>
-                    <option value="oz">Ounces (oz)</option>
-                    <option value="cups">Cups</option>
-                    <option value="grams">Grams (g)</option>
-                    <option value="tsp">Teaspoons (tsp)</option>
-                </select>
+                <div class="static-field">Percentage (%)</div>
+                <input type="hidden" name="unit" value="%">
             </div>
         </div>
 
@@ -56,7 +53,6 @@
                 <textarea id="activity-notes" name="notes" rows="3" style="width: 100%; resize: none;" placeholder="Optional notes about this activity..."></textarea>
             </div>
         </div>
-
 
     </div>
 </form>
