+++ css/components/buttons.css
@@ -15,6 +15,16 @@
     transition: all var(--transition-fast);
 }
 
+.btn:hover {
+  transform: translateY(-3px);
+  box-shadow: 0 8px 20px rgba(0, 0, 0, 0.25);
+}
+
+.btn:active {
+  transform: translateY(-1px);
+  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.2);
+}
+
 /* Small Button */
 .btn-small {
     display: inline-flex;
@@ -29,6 +39,10 @@
     min-width: 75px;
 }
 
+.btn-small:hover {
+  transform: translateY(-1px);
+  box-shadow: 0 8px 20px rgba(0, 0, 0, 0.25);
+}
 
 /* Button Styles */
 /*--------------*/
