+++ css/variables.css
@@ -65,7 +65,7 @@
     --danger-bg: #ffd4c4;
     --danger-text: #8b2500;
     --danger-accent: #d84315;
-    
+
     /* Spacing */
     --spacing-xs: 5px;
     --spacing-sm: 10px;
