+++ index.html
@@ -55,9 +55,17 @@
             <div class="search-container">
                 <input type="text" id="search-input" class="search-input" placeholder="Search...">
             </div>
-            <button class="btn btn-primary" id="new-activity-btn" title="NEW ACTIVITY">
-                + ACTIVITY
-            </button>
+            
+            <!-- buttons section -->
+            <div>
+                <button class="btn btn-primary" id="new-activity-watering-btn" title="ADD WATERING">
+                    + WATERING
+                </button>
+                
+                <button class="btn btn-primary" id="new-activity-btn" title="NEW ACTIVITY">
+                    + ACTIVITY
+                </button>
+            </div>
         </div>
         
         <!-- View container -->
