+++ js/components/toolbar.js
@@ -2,33 +2,36 @@
 // TOOLBAR.JS - Toolbar Component Logic
 // ============================================
 
-import { capitalize } from '../utils.js';
+import { openNewActivityModal } from '../modals/new-activity.controller.js';
+import { openNewWateringModal } from '../modals/activity/watering.controller.js';
+
+/**
+ * Initialize toolbar
+ */
+export function initializeToolbar() {
+    console.log('Toolbar initialized');
+    
+    // Set up buttons event listeners
+    setupButtons();
+}
 
 /**
  * Update toolbar title based on current view
  */
-export function updateToolbarTitle(viewName) {
-    const toolbarTitle = document.getElementById('toolbar-title');
-    const toolbarTagline = document.getElementById('toolbar-tagline');
-    
-    if (!toolbarTitle) return;
-    
-    const viewTitles = {
-        'dashboard': 'Dashboard',
-        'reports': 'Reports',
-        'inventory': 'Inventory',
-        'configuration': 'Configuration',
-        'settings': 'Settings'
-    };
-    
-    const viewTaglines = {
-        'dashboard': 'Your plant care overview',
-        'reports': 'Analytics and insights',
-        'inventory': 'Manage all your plants',
-        'configuration': 'Manage habitat and environmental settings for your plants',
-        'settings': 'Configure your Plant Hub preferences'
-    };
+function setupButtons() {
+    // New Activity button
+    const newActivityBtn = document.getElementById('new-activity-btn');
+    if (newActivityBtn) {
+        newActivityBtn.addEventListener('click', async () => {
+            openNewActivityModal();
+        });
+    }
 
-    toolbarTitle.textContent = viewTitles[viewName] || capitalize(viewName);
-    toolbarTagline.textContent = viewTaglines[viewName] || capitalize(viewName);
+    // New watering button
+    const newWateringBtn = document.getElementById('new-activity-watering-btn');
+    if (newWateringBtn) {
+        newWateringBtn.addEventListener('click', async () => {
+            openNewWateringModal();
+        });
+    }
 }
