+++ js/modals/activity/watering.controller.js
@@ -1,0 +1,58 @@
+// ============================================
+// NEW-ACTIVITY.CONTROLLER.JS - new-activity modal logic
+// ============================================
+
+import { openModal, populateDropdown } from '../prompt-modal.js';
+import { getActivePlants, addPlantActivity } from '../../services/supabase.js';
+import { showNotification } from '../../utils.js';
+
+export async function openNewWateringModal() {
+
+    // Open modal first
+    const modal = await openModal({
+        title: 'Log New Watering',
+        contentUrl: 'components/modals/activity/watering.html',
+        size: 'medium',
+        buttons: [
+            { label: 'LOG ACTIVITY', type: 'primary', action: 'submit' },
+            { label: 'CANCEL', type: 'secondary', action: 'close' }
+        ],
+        onSubmit: async (data, modal) => {
+            try {
+                await addPlantActivity(data);
+                showNotification('Activity logged successfully!', 'success');
+                modal.querySelector('[data-action="close"]').click();
+                this.loadView(this.currentView);
+            } catch (error) {
+                console.error('Error logging activity:', error);
+                showNotification('Failed to log activity', 'error');
+            }
+        }
+    });
+
+    await populateForm(modal);
+}
+
+async function populateForm(modal) {
+    try {
+    // Fetch data from Supabase
+    const [plants] = await Promise.all([
+        getActivePlants()
+    ]);
+    
+    // Populate dropdowns
+    populateDropdown('plant-select', plants, 'plant_id', 'full_plant_name');
+    
+    // Set today's date as default
+    const dateInput = modal.querySelector('#activity-date');
+    if (dateInput) {
+        dateInput.valueAsDate = new Date();
+    }
+    }
+    catch (error) {
+        console.error('Error loading form data:', error);
+        showNotification('Error loading form data', 'error');
+    }
+}
+
+console.log('openModal function:', typeof openModal);
